feat: add day-night cycle analog disk for displaying day or night
